\name{upload_txt}
\alias{upload_txt}
\title{Import and Save Raw Data from N2a Assay TXT Files}
\description{
  Imports and saves raw data generated by the N2a assay stored in `.txt` files with a 96-well plate layout.
}
\usage{
  upload_txt()
}
\details{
  The function reads a `.txt` file selected by the user, which contains multiple datasets separated by blank lines. It imports the datasets into a list, converts any commas to decimal points, and stores the data as numeric values.
}
\value{
  A list of data frames, where each data frame corresponds to a dataset from the `.txt` file. The list is named according to the dataset number, e.g., `dataset_1`, `dataset_2`, etc.
}

\author{
  Synne Thorbjørnsen Frøstrup
}
\note{
  The function assigns the imported datasets to a variable in the global environment with the same name as the file (without the extension).
}


